Split views.py into multiple modules; Add ability to decide where to put Eggsplode when defusing
