+++ messages.json
@@ -7,6 +7,7 @@
   "before_steal": "{} <@{}> wants to steal a card from <@{}>! Accept?",
   "bug_report_print": "===== Bug report from {} in channel {}: [{}] '{}'",
   "bug_reported": "🐞 Bug reported! Thank you for your help!\nYou can also join our support server at https://discord.gg/UGm36FkGDF",
+  "defuse_prompt": "🔧 You drew an Eggsplode card! Please decide where on the deck you want to put it.\n**Current position: {}**\nPlayers:\n{}",
   "defused": "## 🔧 <@{}> drew an Eggsplode card! Luckily, they had a Defuse and put it back into the deck!",
   "eggsploded": "## 💥 <@{}> drew an Eggsplode card and died!",
   "game_already_exists": "❌ Another game is already running in this channel!",
@@ -47,7 +48,7 @@
   "not_implemented": "🙁 Sorry, not implemented yet.",
   "not_your_turn": "❌ It's not your turn!",
   "play_prompt": "**Play** as many cards as you want, then **draw** a card to end your turn!\n## Your cards:\n{}\n-# Cards remaining on the deck: {} | Eggsplode cards: {}",
-  "players_list_item": "{}\n- <@{}>",
+  "players_list_item": "- <@{}>",
   "predicted": "🔮 <@{}> looked at the next 3 cards on the deck!",
   "shuffled": "🌀 <@{}> shuffled the deck!",
   "start": "# New game\n<@{}> wants to start a new Eggsplode game! Click on **Join** to participate!\n**Players:**\n- <@{}>",
